% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_melfcc_small}
\alias{rf_melfcc_small}
\title{Pretrained model
Random forest on Melfcc-transformed data
Butchered with \code{butcher} package to make smaller.}
\format{
A pretrained random forest tidymodels object.
}
\usage{
rf_melfcc_small
}
\description{
Pretrained model
Random forest on Melfcc-transformed data
Butchered with \code{butcher} package to make smaller.
}
\keyword{datasets}
