% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_app_helpers.R
\name{summarize_selection_table}
\alias{summarize_selection_table}
\title{Summarize Selection Table}
\usage{
summarize_selection_table(selection_table)
}
\arguments{
\item{selection_table}{A selection table created by to_selection_table()}
}
\value{
A one-row summary data.frame
}
\description{
Summarize Selection Table
}
