% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_app_helpers.R
\name{to_selection_table}
\alias{to_selection_table}
\title{Helper function for selection table}
\usage{
to_selection_table(df, name, w)
}
\arguments{
\item{df}{A data frame}

\item{name}{The name of the model}

\item{w}{The size of the time window, in seconds}
}
\description{
Helper function for selection table
}
